requirements.txt
fastapi[standard]
pydantic
pytest
pyjwt
passlib[bcrypt]
python-jose[cryptography]